Refactor common code into common header file.
